95% of Subscription-Payment Stuff

+++ Controllers/PlanController.cs
@@ -1,4 +1,5 @@
-﻿using AIDentify.IRepositry;
+﻿using AIDentify.ID_Generator;
+using AIDentify.IRepositry;
 using AIDentify.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -7,14 +8,16 @@ namespace AIDentify.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [Authorize(Roles ="Admin")]
+    //[Authorize(Roles ="Admin")]
     public class PlanController: ControllerBase
     {
         private readonly IPlanRepository PlanRepository;
+        private readonly IdGenerator _idGenerator;
 
-        public PlanController(IPlanRepository planRepository)
+        public PlanController(IPlanRepository planRepository, IdGenerator idGenerator)
         {
             PlanRepository = planRepository;
+            _idGenerator = idGenerator;
         }
 
         [HttpGet]
@@ -32,6 +35,7 @@ namespace AIDentify.Controllers
         [HttpPost]
         public IActionResult Add([FromBody] Plan plan)
         {
+            plan.Id = _idGenerator.GenerateId<Plan>(ModelPrefix.Plan);
             PlanRepository.Add(plan);
             return Ok("Posted Successfully");
         }
@@ -39,12 +43,12 @@ namespace AIDentify.Controllers
         [HttpPut("{id}")]
         public IActionResult Update(string id, [FromBody] Plan plan)
         {
-            if (id != plan.PlanId)
+            if (id != plan.Id)
             {
                 return BadRequest("ID in the URL does not match ID in the body.");
             }
 
-            var existingPlan = PlanRepository.Get(plan.PlanId);
+            var existingPlan = PlanRepository.Get(plan.Id);
             if (existingPlan == null)
             {
                 return NotFound("Plan not found.");
@@ -62,8 +66,15 @@ namespace AIDentify.Controllers
             Plan? plan = PlanRepository.Get(id);
             if(plan != null)
             {
-                PlanRepository.Delete(plan);
-                return Ok("Deleted Successfully");
+                if (PlanRepository.Deleteable(id))
+                {
+                    PlanRepository.Delete(plan);
+                    return Ok("Deleted Successfully");
+                }
+                else
+                {
+                    return BadRequest("Plan has active subscriptions. Cannot delete.");
+                }
             }
             else
             {
